still trying to figure out how to add plot
