+++ package.json
@@ -13,6 +13,7 @@
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/adapter-static": "^1.0.0-next.50",
 		"@sveltejs/kit": "^1.5.0",
+		"@types/d3": "^7.4.2",
 		"svelte": "^3.54.0",
 		"svelte-check": "^3.0.1",
 		"tslib": "^2.4.1",
@@ -21,8 +22,10 @@
 	},
 	"type": "module",
 	"dependencies": {
+		"chart.js": "^4.4.0",
 		"d3": "^7.8.5",
 		"d3-dsv": "^3.0.1",
+		"save-dev": "^0.0.1-security",
 		"simple-statistics": "^7.8.3"
 	}
 }
