+++ src/assets/compared_RQs_sensitivity.csv
@@ -1,0 +1,66 @@
+"city","rq_comm","rq_zip","rq_hdbscan","display_title","region"
+"Bakersfield",1.18308359841345,0.980774812736888,0.945859468847414,"Bakersfield, CA","Pacific"
+"San Jose",1.0755969622137,0.987428955727427,0.964918320398636,"San Jose, CA","Pacific"
+"Wichita",0.952408188921328,0.870307538037098,0.890966783601944,"Wichita, KS","Midwest"
+"Miami",0.928545334015168,0.910486468576513,0.920359627811871,"Miami, FL","Southeast"
+"Los Angeles",0.910935555804358,0.867965999134946,0.831905069522817,"Los Angeles, CA","Pacific"
+"Las Vegas",0.910141812679316,1.12152602043921,1.02636460966033,"Las Vegas, NV","Southwest"
+"Milwaukee",0.905307198644272,0.988468170002813,0.857170402444071,"Milwaukee, WI","Midwest"
+"Vancouver",0.899309391713206,0.790859743710388,0.852328257855662,"Vancouver, BC","Canada"
+"San Diego",0.893547166693029,0.976713504256477,0.796466853644414,"San Diego, CA","Pacific"
+"Salt Lake City",0.889352580257882,0.966512792408613,0.819177751088665,"Salt Lake City, UT","Southwest"
+"Colorado Springs",0.869998399696531,0.874053602603594,0.873607568119905,"Colorado Springs, CO","Southwest"
+"Mississauga",0.83594800560588,0.922684423244434,0.911440825585151,"Mississauga, ON","Canada"
+"Dallas",0.82215503319927,0.762260816994189,0.702780863338522,"Dallas, TX","Southwest"
+"Ottawa",0.821161688929709,0.836294770203094,0.822822023762245,"Ottawa, ON","Canada"
+"Sacramento",0.820812766426167,0.719509836056999,0.657146143816075,"Sacramento, CA","Pacific"
+"Orlando",0.815668633324942,0.918897402051879,0.791095318909765,"Orlando, FL","Southeast"
+"Oklahoma City",0.815284979379306,0.930693729538314,0.929554481855815,"Oklahoma City, OK","Southwest"
+"Tampa",0.808130951520025,0.915306983155711,0.806812320638388,"Tampa, FL","Southeast"
+"Nashville",0.806486460063943,0.829803212040279,0.876595954849634,"Nashville, TN","Southeast"
+"Tucson",0.803250237401865,0.934612134362503,0.911952892467144,"Tucson, AZ","Southwest"
+"Montreal",0.801717767059722,0.860440603860333,0.669858490739191,"Montreal, QC","Canada"
+"Winnipeg",0.801538872648134,0.873986293098836,0.761320122003225,"Winnipeg, MB","Canada"
+"Oakland",0.795924618606653,0.840545871960254,0.743834816934551,"Oakland, CA","Pacific"
+"Boston",0.781461902087815,0.610299369433309,0.702302144198704,"Boston, MA","Northeast"
+"London",0.779525691426814,0.825072701449324,0.790509862054023,"London, ON","Canada"
+"Fort Worth",0.771218720568134,0.913813820036084,0.801332894197183,"Fort Worth, TX","Southwest"
+"El Paso",0.768507430314883,1.14214370242431,0.965238638501394,"El Paso, TX","Southwest"
+"Edmonton",0.761949744972264,0.742759346267296,0.79927761910536,"Edmonton, AB","Canada"
+"Calgary",0.761676000722155,0.849609753665045,0.85187822148427,"Calgary, AB","Canada"
+"New York",0.760034369551418,0.666758251356688,0.658605539412826,"New York, NY","Northeast"
+"San Francisco",0.754418785587669,0.60378237955495,0.673164327421333,"San Francisco, CA","Pacific"
+"Quebec",0.749379558498276,0.834849304409654,0.870927313810999,"Quebec, QC","Canada"
+"Jacksonville",0.746731984614687,0.959214628408293,0.702452590029022,"Jacksonville, FL","Southeast"
+"Austin",0.744620968711382,0.775481836059718,0.726433542391595,"Austin, TX","Southwest"
+"Toronto",0.732199917505211,0.749099858070949,0.697519534031847,"Toronto, ON","Canada"
+"Detroit",0.72453913298383,0.711286112758423,0.658492057083498,"Detroit, MI","Midwest"
+"Indianapolis",0.720460158790306,0.831113593402461,0.793680519557432,"Indianapolis, IN","Midwest"
+"Washington DC",0.716496510200517,0.705903136060821,0.68505703730598,"Washington DC","Northeast"
+"Phoenix",0.704185622144863,0.926850403720882,0.891630832112041,"Phoenix, AZ","Southwest"
+"Halifax",0.702902856624854,0.804715514628524,0.849515942477487,"Halifax, NS","Canada"
+"Chicago",0.701135981918199,0.612941650657934,0.614576625554533,"Chicago, IL","Midwest"
+"Baltimore",0.699006370042386,0.797767374392683,0.677724457482987,"Baltimore, MD","Northeast"
+"San Antonio",0.687035593750934,0.82351998083594,0.758478141447656,"San Antonio, TX","Southwest"
+"Tulsa",0.685573391421713,0.83430118194947,0.821499104708219,"Tulsa, OK","Southwest"
+"Denver",0.683393081740602,0.708027138648698,0.66794486022482,"Denver, CO","Southwest"
+"Cincinnati",0.67416022801258,0.753363639580359,0.644403997291466,"Cincinnati, OH","Midwest"
+"Pittsburgh",0.673909620364597,0.774942289037317,0.662843889177992,"Pittsburgh, PA","Northeast"
+"Fresno",0.670720168919165,0.776116103271743,0.71832652409348,"Fresno, CA","Pacific"
+"Kansas City",0.665938181608569,0.800580199287844,0.738697644141633,"Kansas City, MO","Midwest"
+"Cleveland",0.662719502238278,0.82701357860519,0.781381481918943,"Cleveland, OH","Midwest"
+"Houston",0.656492842248641,0.780319331145923,0.639464348191383,"Houston, TX","Southwest"
+"Portland",0.655381106536047,0.510761719524214,0.60787214632729,"Portland, OR","Pacific"
+"Philadelphia",0.63798789706899,0.688932126230554,0.672484329496864,"Philadelphia, PA","Northeast"
+"Charlotte",0.633754472206764,0.847648601382973,0.745151012339772,"Charlotte, NC","Southeast"
+"Memphis",0.629771677133442,0.821322341196049,0.739803598655364,"Memphis, TN","Southeast"
+"Albuquerque",0.629474230277162,0.907255071359103,0.662453049462028,"Albuquerque, NM","Southwest"
+"Seattle",0.623177244903676,0.669363699738344,0.56507236523636,"Seattle, WA","Pacific"
+"Raleigh",0.618937996629154,0.720619117510862,0.632144242390279,"Raleigh, NC","Southeast"
+"Louisville",0.602049871120794,0.757390467386113,0.542219286509983,"Louisville, KY","Midwest"
+"Atlanta",0.597757299863599,0.847798764858894,0.852568124208396,"Atlanta, GA","Southeast"
+"Omaha",0.576850060860936,0.796789214051663,0.880426022227232,"Omaha, NE","Midwest"
+"New Orleans",0.559795411379932,0.793556189759655,0.744811405443444,"New Orleans, LA","Southeast"
+"St Louis",0.545597388527244,0.58834826492473,0.53158900009084,"St. Louis, MO","Midwest"
+"Columbus",0.452794829227609,0.722441548959234,0.565491365183843,"Columbus, OH","Midwest"
+"Minneapolis",0.444687605804483,0.547229575677818,0.555528987671666,"Minneapolis, MN","Midwest"
