+++ src/routes/blog/sensitivity-analysis/lib/Map.svelte
@@ -4,6 +4,7 @@
 	import maplibregl from 'maplibre-gl';
 	import hdbscan_downtowns from '../../../../assets/HDBSCAN_downtowns_with_province.geo.json';
 	import old_downtowns from '../../../../assets/old_study_areas.geo.json';
+	import { cityCoordinates } from './city_coords.js';
 
 	let pageHeight;
 	let pageWidth;
@@ -11,6 +12,7 @@
      * @type {import("maplibre-gl").Map}
      */
 	let map;
+	let selectedCity = 'Albuquerque NM'; // Default city
 
 	let showHDBSCANDowntowns = true; // Initial visibility
 	let showOldDowntowns = true; // Initial visibility
@@ -26,7 +28,7 @@
 
 		map = new maplibregl.Map({
 			container: 'map',
-			center: [-79.386, 43.653], 
+			center: [cityCoordinates[selectedCity].long, cityCoordinates[selectedCity].lat],
 			zoom: 12,
 			minZoom: 9,
 			maxZoom: 16,
@@ -38,7 +40,6 @@
 
 		map.dragRotate.disable();
 		map.touchZoomRotate.disableRotation();
-		// map.scrollZoom.disable();
 
 		map.addSource('hdbscan_downtowns', {
 			'type': 'geojson',
@@ -74,8 +75,8 @@
 			'type': 'raster',
 			'tiles': ['https://tile.openstreetmap.org/{z}/{x}/{y}.png'],
 			'tileSize': 256,
-			'minzoom': 0,
-			'maxzoom': 19
+			// 'minzoom': 0,
+			// 'maxzoom': 19
 		});
 		map.addLayer({
 			'id': 'osm-raster-tiles',
@@ -97,51 +98,80 @@
     	}
 	}	
 
+	function updateMap() {
+		// Update the map's center and zoom based on the selected city
+		const { lat, long } = cityCoordinates[selectedCity];
+		map.flyTo({ center: [long, lat], zoom: 12, duration: 50 }); // Adjust the zoom level as needed
+  	}	
+
 </script>
 
 <svelte:window bind:innerHeight={pageHeight} bind:innerWidth={pageWidth}/>
 
-<div id="map" style="height: {mapHeight}px"></div>
+<div class="container">
+  <div id="map" style="height: {mapHeight}px"></div>
 
-<div id="bottombar">
+  <div>
 	<div>
-		<label>
-		<input type="checkbox" bind:checked={showHDBSCANDowntowns} />
-		HDBSCAN
-		</label>
-	</div>
-	<div>
-		<label>
-		<input type="checkbox" bind:checked={showOldDowntowns} />
-		Zip code (former definition)
-		</label>
-	</div>
-	<p id='data_sources'>Data Sources: OpenStreetMap</p>
+		<label for="cityDropdown">Select a city:</label>
+		<select id="cityDropdown" bind:value={selectedCity} on:change={updateMap}>
+		  {#each Object.keys(cityCoordinates) as city}
+			<option value={city}>{city}</option>
+		  {/each}
+		</select>
+	  </div>
+	
+	  <div id="sidebar">
+		<div>
+			<label>
+			<input type="checkbox" bind:checked={showHDBSCANDowntowns} />
+			HDBSCAN
+			</label>
+		</div>
+		<br>
+		<div>
+			<label>
+			<input type="checkbox" bind:checked={showOldDowntowns} />
+			Zip code (former definition)
+			</label>
+		</div>
+		<div>
+			<br><br>
+			<p id='data_sources'>Data Sources:<br>- OpenStreetMap<br>- LEHD</p>
+		</div>
+	  </div>
+  </div>
+
 </div>
 
-<!-- <p>Data Sources: OpenStreetMap</p> -->
-
 <style>
-	#map {
-		width: 100%;
-		margin: 0 auto;
-		max-width: 1200px;
-		/* border-top: 1px solid var(--brandBlack);
-		border-bottom: 1px solid var(--brandBlack); */
-		background-color: white;
-	}
+  .container {
+    display: flex;
+    flex-direction: row; /* Horizontal layout */
+    justify-content: space-between; /* Align items to the start and end of the container */
+    align-items: flex-start; /* Align items to the top of the container */
+    max-width: 1200px;
+    margin: 0 auto;
+  }
 
-	#bottombar {
-		margin-left: 20px;
-		margin-right: 10px;
-	}
+  #map {
+	width: 80%;
+	height: 100%;
+	margin: 0 auto;
+	max-width: 900px;
+	background-color: white;	
+  }
 
-	#data_sources {
-		margin: 0 auto;
-		text-align: right;
-		font-size: 12px;
-		max-width: 1200px;
-		color: white;		
-	}
-	
+  #sidebar {
+    width: 30%;
+	margin-left: 20px;
+  }
+
+  #data_sources {
+	margin: 0 auto;
+	text-align: left;
+	font-size: 12px;
+	max-width: 1200px;
+	color: white;		
+  }
 </style>
